+++ src/main/java/org/manuel/mysportfolio/model/dtos/team/UrlImageDto.java
@@ -1,16 +1,18 @@
 package org.manuel.mysportfolio.model.dtos.team;
 
+import javax.annotation.concurrent.Immutable;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
-@lombok.Data
+@Immutable
 @lombok.AllArgsConstructor
-@lombok.NoArgsConstructor
+@lombok.Data
+@lombok.extern.jackson.Jacksonized @lombok.Builder(toBuilder = true)
 public class UrlImageDto implements TeamImageDto {
 
   @org.hibernate.validator.constraints.URL
   @Size(max = 200)
   @NotNull
-  private String url;
+  private final String url;
 
 }